Batch of correctness, stability, and quality fixes
VM correctness / DoS prevention:
- harn-vm/stdlib/json: cap JSON_PARSE_CACHE at 128 entries so a script
  feeding varied JSON cannot grow the per-thread cache without bound
  (mirrors the regex-cache ceiling already in place)
- harn-vm/stdlib/concurrency: positive_u32_arg was casting i64 to u32
  via as, silently wrapping any value above u32::MAX; switch to
  try_from and surface a clear error. Adds conformance test
  sync_semaphore_capacity_overflow that exercises the new error path.
- harn-vm/agent_sessions: close() now drains both the cross-thread
  agent_inbox and event sinks. Previously the inbox was never cleared,
  so a session reusing an old id could receive stale notifications.
  close_with_status() inherits this via close(). Adds regression test.

Test/CI stability - eliminate wall-clock-based uniqueness:
- harn-cli/check tests: unique_temp_dir was seeded with
  SystemTime::now().as_nanos(), which can collide on loaded CI runners
  when two callers land in the same tick. Replace with
  PID + AtomicU64 counter.
- harn-hostlib/run_test: same fix for junit_temp_path; PID alone was
  insufficient when one process called it multiple times per
  nanosecond.

Formatter pretty-print:
- harn-fmt: MethodCall / OptionalMethodCall passed obj.len() +
  method.len() + N as prefix_len to the wrap check, but when obj had
  wrapped to multiple lines that included intermediate newlines and
  indent - over-counting the column position and forcing args to wrap
  unnecessarily. Compute the correct column using a new
  last_line_width helper. Regression tests cover both forms.

Portal frontend:
- DlqPage: window.setInterval for job polling was never cleared on
  unmount; track the timer in a ref and clear it from a useEffect
  cleanup. Also clears any previous poller when a new job replaces it.

Modularity:
- agent_sessions.rs: extract the ~490-line inline #[cfg(test)] mod tests
  into a sibling agent_sessions_tests.rs via the existing
  #[path = "..."] pattern used elsewhere in the crate.
