+++ conformance/tests/errors_tests/sync_semaphore_capacity_overflow.error
@@ -1,0 +1,1 @@
+sync_semaphore_acquire: value 4294967296 exceeds u32::MAX
